refactor: 이벤트 발행 코드 최적화

+++ funding/src/main/java/com/nowayback/funding/application/fundingProjectStatistics/dto/event/ProjectFundingSuccessEvent.java
@@ -37,7 +37,7 @@ public record ProjectFundingSuccessEvent(
     public static ProjectFundingSuccessEvent of(
             UUID projectId,
             Long finalAmount,
-            Long participantCount
+            Integer participantCount
     ) {
         return new ProjectFundingSuccessEvent(
                 UUID.randomUUID(),
